+++ R/prettifyLATEX.R
@@ -1,54 +1,45 @@
-#' Prettify C, C++, Java
-#' @description Prettify some C, C++ or Java code.
+#' Prettify LaTeX
+#' @description Prettify LaTeX code, including Sweave code, \code{sty} files,
+#'   \code{cls} files, and \code{bib} files.
 #'
 #' @param contents the code to be prettified; there are three possibilities for
 #'   this argument:
 #'   \code{NA} (default), to use the file currently opened in RStudio;
 #'   the path to a file;
 #'   or the code given as a character vector
-#' @param language the language of the code; when the contents is read from a
-#'   file, this option is ignored, because the language is obtained from the
-#'   extension of the file
 #' @param tabSize number of spaces of the indentation (usually \code{2} or
 #'   \code{4});
 #'   if \code{NULL} (the default), there are two possibilities:
 #'   if the contents is read from the current file in RStudio, then the
 #'   number of spaces will be the one you use in RStudio;
 #'   otherwise it is set to \code{2}
+#' @param log logical, whether to generate a log file (it will be named
+#'   \code{indent.log})
 #'
 #' @return The pretty code in a character string.
 #'
-#' @note This function requires the command line utility \code{clang-format}.
+#' @note This function requires the command line utility \code{latexindent}.
 #'
 #' @importFrom rstudioapi isAvailable
 #' @importFrom tools file_ext
 #' @export
-prettifyCLANG <- function(contents = NA, language = NA, tabSize = NULL){
+prettifyLaTeX <- function(contents = NA, tabSize = NULL, log = FALSE){
 
-  if(Sys.which("clang-format") == ""){
+  if(Sys.which("latexindent") == ""){
     stop(
-      "This function requires `clang-format`. ",
-      "Either it is not installed, or it is not found. ",
-      "To reindent C/C++/Java code, you can use the 'Shiny Indent' addin."
+      "This function requires `latexindent`. ",
+      "Either it is not installed, or it is not found."
     )
   }
 
-  language <- tolower(language)
-
   if(!isNA(contents) && !isFile(contents)){
-    if(is.na(language)){
-      stop("Please specify the language of this code.")
-    }
-    if(!is.element(language, c("c", "c++", "cpp", "java"))){
-      stop('The language must be one of "c", "c++", or "java".')
-    }
-    ext <- language
+    ext <- "tex" # this works for bib, etc.
   }
   if(isNA(contents) && isAvailable()){
     context <- RStudioContext()
     ext <- tolower(file_ext(context[["path"]]))
-    if(!is.element(ext, c("c", "cpp", "c++", "h", "hpp", "java"))){
-      stop("Looks like this file is not a C/C++/Java file.")
+    if(!is.element(ext, c("tex", "latex", "rnw", "bib", "sty", "cls"))){
+      stop("Looks like this file is not a LaTeX file.")
     }
     if(is.null(tabSize)){
       tabSize <- RStudioTabSize()
@@ -67,18 +58,33 @@ prettifyCLANG <- function(contents = NA, language = NA, tabSize = NULL){
     }
     if(isFile(contents)){
       ext <- tolower(file_ext(contents))
-      if(!is.element(ext, c("c", "cpp", "c++", "h", "hpp", "java"))){
-        stop("Looks like this file is not a C/C++/Java file.")
+      if(!is.element(ext, c("tex", "latex", "rnw", "bib", "sty", "cls"))){
+        stop("Looks like this file is not a LaTeX file.")
       }
       contents <- suppressWarnings(readLines(contents))
     }
   }
+  ext0 <- ifelse(ext == "rnw", "tex", ext)
   tmpDir <- tempdir()
-  writeLines(.clangFormat(tabSize), file.path(tmpDir, ".clang-format"))
-  tmpFile <- tempfile(fileext = paste0(".", ext))
+  tmpFile <- tempfile(fileext = paste0(".", ext0))
   writeLines(contents, tmpFile)
+  cruft <- if(!log) paste0("--cruft=", shQuote(tmpDir))
   prettyCode <- suppressWarnings(system2(
-    "clang-format", tmpFile,
+    "latexindent",
+    c(
+      tmpFile,
+      sprintf(
+        sprintf("-y='%s'",
+                paste0(c(
+                  "defaultIndent:\"%s\"",
+                  "indentRules:displayMath:\"\"",
+                  "indentRules:displayMathTeX:\"\""
+                ), collapse = ",")
+        ),
+        paste0(rep(" ", tabSize), collapse = "")
+      ),
+      cruft
+    ),
     stdout = TRUE, stderr = TRUE
   ))
   if(!is.null(attr(prettyCode, "status"))){