+++ inst/shinyApp/server.R
@@ -1,13 +1,7 @@
-library(rstudioapi)
-
-context <- getOption("prettify.context")
 language <- getOption("prettify.language")
 code <- getOption("prettify.code")
-id <- context[["id"]]
 codemirror <- getOption("prettify.codemirror")
-
 action <- getOption("prettify.action")
-
 tabSize <- getOption("prettify.tabSize")
 
 if(action == "prettify"){
@@ -19,7 +13,8 @@ if(action == "prettify"){
     javascript = "babel",
     jsx = "babel",
     markdown = "markdown",
-    scss = "css"
+    scss = "css",
+    yaml = "yaml"
   )
 }else{
   if(codemirror){
@@ -45,8 +40,7 @@ shinyServer(function(input, output, session){
   })
 
   observeEvent(input[["done"]], {
-    setDocumentContents(text = input[["prettyCode"]], id = id)
-    stopApp()
+    stopApp(input[["prettyCode"]])
   })
 
   session$sendCustomMessage(