+++ inst/shinyApp/www/autoFormatRange.js
@@ -1,7 +1,9 @@
-function indentify(codeAndParser) {
+function indentify(codeAndParserAndTabSize) {
   var editor = CodeMirror.fromTextArea(document.getElementById("editor"), {
+    mode: codeAndParserAndTabSize.parser,
+    tabSize: codeAndParserAndTabSize.tabSize,
+    indentUnit: codeAndParserAndTabSize.tabSize,
     lineNumbers: true,
-    mode: codeAndParser.parser,
     lineWrapping: true,
     theme: theme
   });