+++ NEWS.md
@@ -1,3 +1,23 @@
+# prettifyAddins 2.0.0
+
+* Added a YAML prettifier. It obviously acts on YAML files, but also on the 
+YAML metadata of (R)Markdown files.
+* Added a LaTeX prettifier. It can be used for Sweave files. It requires the 
+command-line utility `latexindent`, which is included in MikTeX and TeX Live.
+* Added a Julia prettifier. It requires Julia and the Julia package 
+`JuliaFormatter`.
+* Added a Python prettifier. It requires [black](https://github.com/psf/black).
+* Added another Java prettifier, a JSON prettifier and a Ruby prettifier 
+(function `prettify_FCA`).
+* Now the package provides some functions to prettify, which return the pretty 
+code. The addins just call these functions and render the pretty code in 
+RStudio.
+* Among these functions, there is `reindent_PhantomJS`. It allows to reindent 
+code for some languages which were supported by a Shiny addin only, whereas 
+this function does not run a Shiny app. These languages are C, C++, Java, 
+Fortran, Julia, Python, SAS, Scala, Shell, and SQL.
+
+
 # prettifyAddins 1.0.0
 
 * Added non-interactive versions of the addins, thanks to the V8 package.